feat: implement create and update functionality for modalities with new dialog forms
